analytics_api/requirements.txt
fastapi==0.115.0
uvicorn==0.32.0
pydantic==2.9.2
pydantic-settings==2.6.0
motor==3.6.0
pymongo>=4.9,<4.10
python-jose[cryptography]==3.3.0
pyjwt==2.9.0
httpx==0.27.2
python-dotenv==1.0.0
